requirements.txt
python==3.5
tensorflow==1.5
keras=1.1.0
flask
torch
torchvision
pillow
